+++ src/main/java/com/proyecto/terranova/implement/ProductoImplement.java
@@ -1,5 +1,8 @@
 package com.proyecto.terranova.implement;
 
+import com.proyecto.terranova.entity.Usuario;
+import com.proyecto.terranova.repository.UsuarioRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.util.List;
@@ -16,52 +19,54 @@ public class ProductoImplement implements ProductoService {
     @Autowired
     private ProductoRepository repository;
 
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+
+    @Autowired
+    private ModelMapper modelMapper;
+
     @Override
     public ProductoDTO save(ProductoDTO dto) {
-        Producto entity = new Producto();
-        // TODO: map DTO to Entity
-        entity = repository.save(entity);
-        // TODO: map Entity to DTO
-        return dto;
+        Producto entidadProducto = modelMapper.map(dto, Producto.class);
+        Usuario usuario = usuarioRepository.findById(dto.getCedulaVendedor())
+                .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+        entidadProducto.setUsuario(usuario);
+        Producto entidadGuardada = repository.save(entidadProducto);
+        return modelMapper.map(entidadGuardada, ProductoDTO.class);
     }
 
     @Override
     public ProductoDTO update(Long id, ProductoDTO dto) {
-        return repository.findById(id)
-            .map(entity -> {
-                // TODO: actualizar campos desde DTO a entity
-                entity = repository.save(entity);
-                // TODO: map Entity to DTO
-                return dto;
-            })
-            .orElse(null);
+        Producto entidadProducto = repository.findById(id)
+        .orElseThrow(() -> new RuntimeException("Producto no encontrado"));
+
+    	modelMapper.map(dto, entidadProducto);
+
+    	Producto entidadActualizada = repository.save(entidadProducto);
+    	return modelMapper.map(entidadActualizada, ProductoDTO.class);
     }
 
     @Override
     public ProductoDTO findById(Long id) {
-        return repository.findById(id)
-            .map(entity -> {
-                ProductoDTO dto = new ProductoDTO();
-                // TODO: map Entity to DTO
-                return dto;
-            })
-            .orElse(null);
+        Producto entidadProducto = repository.findById(id).orElseThrow(() -> new RuntimeException("Producto no encontrado"));
+        return modelMapper.map(entidadProducto, ProductoDTO.class);
     }
 
     @Override
     public List<ProductoDTO> findAll() {
         return repository.findAll().stream()
-            .map(entity -> {
-                ProductoDTO dto = new ProductoDTO();
-                // TODO: map Entity to DTO
-                return dto;
-            })
+            .map(entity -> modelMapper.map(entity, ProductoDTO.class))
             .collect(Collectors.toList());
     }
 
     @Override
-    public void delete(Long id) {
+    public boolean delete(Long id) {
+        if(!repository.existsById(id)){
+               return false;
+        }
         repository.deleteById(id);
+        return true;
     }
 
     @Override
